app/requirements.txt
python-dotenv
asgiref==3.2.10
Django==3.1.2
gunicorn==20.0.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-material-admin==1.7.15
djangorestframework==3.12.1
django-rest-knox==4.1.0
django-widget-tweaks==1.4.8
future==0.18.2
geojson==2.5.0
Markdown==3.3.3
numpy==1.19.3
pandas==1.1.4
#psycopg2-binary==2.8.5
psycopg2
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
topojson
uk-postcode-utils==1.1
xlrd==1.2.0
django-tinymce
django-json-widget
django-referrer-policy
django-guardian
django-jsonform
django-localflavor==2.1
trafilatura
area